NominusBeta/src/test/java/com/beta/qa/testcases/LoginPageTest.java
package com.beta.qa.testcases;

import static org.testng.Assert.assertEquals;

import org.testng.Assert;
import org.testng.annotations.AfterMethod;
import org.testng.annotations.BeforeMethod;
import org.testng.annotations.Test;

import com.beta.qa.base.TestBase;
import com.beta.qa.pages.HomePage;
import com.beta.qa.pages.LoginPage;

public class LoginPageTest extends TestBase {
	LoginPage loginPage;
	HomePage homePage;

	public LoginPageTest() {
		super();
	}

	@BeforeMethod
	public void setup() {
		initialization();
		loginPage = new LoginPage();
	}

	@Test(priority = 1)
	public void loginPageTitleTest() {

		String title = loginPage.validateLoginPageTitle();
		Assert.assertEquals(title, "Domain & Trademark Registration Services");

	}

	@Test(priority = 2)
	public void NominusLogoTest() {
		boolean flag = loginPage.validateNominusLogo();
		Assert.assertTrue(flag);
	}

	@Test(priority = 3)
	public void loginTest() {
		homePage = loginPage.Login(prop.getProperty("username"), prop.getProperty("password"));
	}

	@AfterMethod
	public void tearDown() {
		driver.quit();
	}

}
